[FIX] UI bugs & data-platform backend fixes (#463)
* fix(ui): fix potential page crash bug

* fix(data-discovery): fix bug where upstream/downstream tasks do not match corresponding task definition

* fix(ui): display task run title on navigator breadcrumb

* fix(data-platform): fix notification display on upstream tasks not deployed & fix HTTP code to 409 when throwing `IllegalStateException`

* fix(data-platform): use user defined exception class `UpstreamTaskNotPublisedException` instead of `IllegalStateException`

* wip: task run service log reading

* feat(task-run-log): enhance task run log API & add limitation to polling log viewer component

* feat(ui): display properties of taskruns on DAG panel

* fix(ui): redirect to dashboard page by default

* chore(ui): remove unused "recently viewed" section

* test(workflow): fix workflow test cases

* fix(workflow): fix variable service test

* fix: total count field should be an integer

* fix(data-platform): sort query result in id descending order (latest deployed tasks comes first)

* fix(data-dashboard): filter out non-scheduled taskruns (dry runs)

* fix: typo

* fix(data-discovery): filter out non-scheduled upstream/downstream task affection on lineage

* test(data-platform): fix test cases

* fix(data-platform): fix deserialization bug on class `TaskNotifyConfig`

* feat(task-definition): support schedule type "manual" (ScheduleType.NONE)

(cherry picked from commit 02f9709880c70c321ce928e2193c0589442244f8)

* feat(task-definition): improve cron expression editor

(cherry picked from commit f8d4305459cd8ae515e42b64c3579a5bcec400f8)

* feat(ui): change icon of abort/restart & add "scheduled time" column in scheduled task runs page

* fix(dashboard): fix pending tasks count always be zero

* fix(ui): colored status text & fix table layout

* feat(scheduled-tasks): allow to view historical task attempts information in scheduled tasks table

* chore: remove `console.log`

* feat(ui): allow stopping dry run & display dry run task status

* fix: scheduled tasks page should have correct breadcrumb title

* fix(ui): fix display logic on end time

* fix(data-platform): do validation on configured Cron expression before committing task snapshot to workflow

* fix: build dependency

* feat(ui): support new form type: "single-select" and "list"

+++ kun-data-dashboard/src/main/java/com/miotech/kun/datadashboard/model/entity/DataDevelopmentMetrics.java
@@ -9,15 +9,15 @@ import lombok.Data;
 @Data
 public class DataDevelopmentMetrics {
 
-    Long successTaskCount;
+    Integer successTaskCount;
 
-    Long failedTaskCount;
+    Integer failedTaskCount;
 
-    Long runningTaskCount;
+    Integer runningTaskCount;
 
-    Long startedTaskCount;
+    Integer startedTaskCount;
 
-    Long pendingTaskCount;
+    Integer pendingTaskCount;
 
-    Long totalTaskCount;
+    Integer totalTaskCount;
 }
